Have the HttpResponseException filter use the new pattern
